+++ src/rules/no-extra-semicolons/__tests__/index.js
@@ -37,6 +37,10 @@ testRule(rule, {
     code: "a { content: ';;'; }",
   }, {
     code: "a { content: ';\t; ;'; }",
+  }, {
+    code: ":root { --foo: red; --bar: blue; }",
+  }, {
+    code: ":root { --foo: { color: red }; --bar: { color: blue }; --foo-bar: { color: blue }; --bar-foo: { color: red }; }",
   } ],
 
   reject: [ {
@@ -474,6 +478,41 @@ testRule(rule, {
     message: messages.rejected,
     line: 6,
     column: 1,
+  }, {
+    code: ":root { --foo: red;; --bar: blue; }",
+    message: messages.rejected,
+    line: 1,
+    column: 20,
+  }, {
+    code: ":root { --foo: red; --bar: blue;; }",
+    message: messages.rejected,
+    line: 1,
+    column: 33,
+  }, {
+    code: ":root {; --foo: { color: red }; --bar: { color: blue }; --bar-foo: { color: red }; }",
+    message: messages.rejected,
+    line: 1,
+    column: 8,
+  }, {
+    code: ":root { --foo: { color: red };; --bar: { color: blue }; --bar-foo: { color: red }; }",
+    message: messages.rejected,
+    line: 1,
+    column: 31,
+  }, {
+    code: ":root { --foo: { color: red }; --bar: { color: blue };; --bar-foo: { color: red }; }",
+    message: messages.rejected,
+    line: 1,
+    column: 55,
+  }, {
+    code: ":root { --foo: { color: red }; --bar: { color: blue }; --bar-foo: { color: red };; }",
+    message: messages.rejected,
+    line: 1,
+    column: 82,
+  }, {
+    code: "a { --foo: { color: red }; --bar: { color: blue }; --bar-foo: { color: red };; }",
+    message: messages.rejected,
+    line: 1,
+    column: 78,
   } ],
 })