+++ src/utils/__tests__/isStandardSyntaxDeclaration-test.js
@@ -5,8 +5,7 @@ import scss from "postcss-scss"
 import test from "tape"
 
 test("isStandardSyntaxDeclaration", t => {
-
-  t.plan(18)
+  t.plan(22)
 
   rules("a { a: b }", decl => {
     t.ok(isStandardSyntaxDeclaration(decl), "standard prop and value")
@@ -20,6 +19,9 @@ test("isStandardSyntaxDeclaration", t => {
   rules("a { a : calc(b + c) }", decl => {
     t.ok(isStandardSyntaxDeclaration(decl), "standard prop and calc value")
   })
+  rules("@page { size: A4 }", decl => {
+    t.ok(isStandardSyntaxDeclaration(decl), "does not break @selector")
+  })
   scssRules("a { #{$var}: 10px; }", decl => {
     t.ok(isStandardSyntaxDeclaration(decl), "property with scss variable interpolation (only)")
   })
@@ -28,6 +30,9 @@ test("isStandardSyntaxDeclaration", t => {
   })
   scssRules("a { prop#{$var}erty: 10px; }", decl => {
     t.ok(isStandardSyntaxDeclaration(decl), "property with scss variable interpolation (middle)")
+  })
+  scssRules("a { --custom-property-set: { color: blue; } }", decl => {
+    t.ok(isStandardSyntaxDeclaration(decl), "custom property set")
   })
   lessRules("a { @{var}: 10px; }", decl => {
     t.ok(isStandardSyntaxDeclaration(decl), "property with less variable interpolation (only)")
@@ -53,6 +58,9 @@ test("isStandardSyntaxDeclaration", t => {
   })
   scssRules("a { $list: (key: value, key2: value2) }", decl => {
     t.notOk(isStandardSyntaxDeclaration(decl), "nested scss list")
+  })
+  scssRules("a { border: { style: solid; color: red; } }", decl => {
+    t.notOk(isStandardSyntaxDeclaration(decl), "scss nested property")
   })
   scssRules("a { $map: (value, value2) }", decl => {
     t.notOk(isStandardSyntaxDeclaration(decl), "nested scss map")