+++ CHANGELOG.md
@@ -6,10 +6,21 @@
 -   Added: if `resolveFrom` fail and `lookup` is id, try resolve from `process.cwd()`
 -   Added: extend configs can be a function now.
 -   Added: API for configuration file/object. `stylelint.buildConfig`
+
+# 7.3.1
+
+-   Fixed: regression in 7.3.0 which caused a "Cannot read property 'length' of undefined" error on a certain selector.
+
+# 7.3.0
+
+-   Added: `processors` can accept options objects.
 -   Added: `ignore: ["inside-function"]` option to `color-named`.
--   Fixed: `block-no-single-line` now ignores empty blocks.
 -   Fixed: `--ignore-path` and `--report-needless-disables` no longer fails when used together.
--   Fixed: the `indentation` rule now correctly handles `_` hacks on property names.
+-   Fixed: `block-closing-brace-newline-after` and `block-closing-brace-space-after` now allow a trailing semicolon after the closing brace of a block.
+-   Fixed: `block-no-single-line` now ignores empty blocks.
+-   Fixed: `declaration-block-no-ignored-properties` now uses the value of the last occurrence of a triggering property.
+-   Fixed: `indentation` now correctly handles `_` hacks on property names.
+-   Fixed: `property-no-unknown` now ignores SCSS nested properties.
 
 # 7.2.0